% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%any_like\%}
\alias{\%any_like\%}
\alias{\%all_like\%}
\alias{\%any_in\%}
\alias{\%all_in\%}
\title{Match value lists against patterns}
\usage{
lhs \%any_like\% rhs

lhs \%all_like\% rhs

lhs \%any_in\% rhs

lhs \%all_in\% rhs
}
\arguments{
\item{lhs}{a character/integer vector or a list of character/integer vectors.
Make sure all list elements have the same type as \code{rhs}}

\item{rhs}{a character/integer vector. In case of a character vector,
wildcard values \code{\%} can be used

It is recommended that all strings only contain ASCII characters.}
}
\value{
A logical vector with the same length as 'lhs'
}
\description{
Checks if any|all elements on the left hand side are matched exactly by one
of the patterns on the right hand side.
}
\examples{
list("abcde", c("bcd", "abc")) \%any_like\% "ab\%"

list("abcde", c("bcd", "abc")) \%all_like\% "ab\%"

list("abcde", c("bcd", "abc")) \%any_in\% c("abc", "abcde")

list("abcde", c("bcd", "abc")) \%all_in\% c("abc", "abcde")

}
