% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%==\%}
\alias{\%==\%}
\alias{\%!=\%}
\alias{\%>=\%}
\alias{\%<=\%}
\alias{\%<\%}
\alias{\%>\%}
\title{Binary comparison operators}
\usage{
lhs \%==\% rhs

lhs \%!=\% rhs

lhs \%>=\% rhs

lhs \%<=\% rhs

lhs \%<\% rhs

lhs \%>\% rhs
}
\arguments{
\item{lhs}{the left hand side of the operator}

\item{rhs}{the right hand side of the operator}
}
\description{
Operators that perform vectorized comparisons, treating \code{NA}s as
\code{FALSE}.
}
\details{
However, equality tests between two \code{NA} values yield \code{TRUE}.
In that case a warning is issued.
}
\examples{
2 \%==\% 2

2 \%==\% 3

2 \%==\% NA

NA \%==\% NA

}
