% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%between\%}
\alias{\%between\%}
\title{Between}
\usage{
lhs \%between\% rhs
}
\arguments{
\item{lhs}{left hand side}

\item{rhs}{right hand side, needs to be a numeric vector of length 2 with no
NAs.}
}
\value{
logical vector of length `length(lhs)`
}
\description{
Checks if a value is between two values (limits included). It treats
\code{NA} as \code{FALSE}.
}
