% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-by.R
\name{\%group_by\%}
\alias{\%group_by\%}
\title{Aggregate values within groups}
\usage{
lhs \%group_by\% rhs
}
\arguments{
\item{lhs}{an aggregation expression}

\item{rhs}{a single name or a vector of names, where 'name' refers to the
  data type \code{name}, i.e. a quoted symbol.

The function dynamically constructs a data.frame based on the values that are
used within the two expressions. It is assumed that all symbols in the
expression refer to character or numeric vectors of the same length in the
calling environment.}
}
\description{
Aggregate values within groups
}
\examples{
with(mtcars, max(hp) \%group_by\% cyl)

with(mtcars, max(hp) \%group_by\% c(cyl, drat))

}
