% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqtigfunctions-package.R
\docType{package}
\name{iqtigfunctions-package}
\alias{iqtigfunctions}
\alias{iqtigfunctions-package}
\title{iqtigfunctions: IQTIG Specific Functions for Public Reporting}
\description{
The goal of iqtig.lang is to provide an implementation of IQTIG specific R
functions as e.g. used in the published indicator specifications (QIDB). For a
detailed, technical reference please use the `help` function, or take a look
at the IQTIG R readme (in German), distributed with the QIDB.
}
\details{
Please also take note of \code{vignette("qskh_qidb", package =
"iqtigfunctions")}, which showcases how you can import a QSKH
QIDB Microsoft Access Database and apply the computational rules of various
indicators to a dataset.

For any feedback, ideas or bug reports, please do not hesitate to contact the
maintainer.
}
\author{
\strong{Maintainer}: IQTIG Verfahrenssupport \email{verfahrenssupport@iqtig.org}

Authors:
\itemize{
  \item Institut für Qualitätssicherung und Transparenz im Gesundheitswesen (IQTIG) [copyright holder]
}

}
