% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{minimum}
\alias{minimum}
\alias{maximum}
\title{Returns the minimum or maximum of a vector}
\usage{
minimum(x)

maximum(x)
}
\arguments{
\item{x}{a vector}
}
\description{
If the sequence is empty, it returns a type stable NA. NAs are ignored.
}
\examples{
minimum(1:10)
minimum(NA)
}
