% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NAs}
\usage{
replace_na(x, by)
}
\arguments{
\item{x}{a vector of values with potential NAs}

\item{by}{a replacement value of the same type as x}
}
\value{
A vector of the same length as x with NAs replaced by 'by'
}
\description{
Replaces all NAs in the first argument by the value in the second argument.
}
\details{
Fails if both arguments do not have the same type.
}
\examples{
replace_na(c(20, NA, 10), 15)

}
