% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{row_sums}
\alias{row_sums}
\title{Sums up multiple variables by row}
\usage{
row_sums(...)
}
\arguments{
\item{...}{an arbitrary number of numeric or logical 1D vectors}
}
\value{
a numeric vector >= 0
}
\description{
Similar to \code{\link{rowSums}}, but treats NAs as 0 and is intented to
accept multiple vectors.
}
\examples{

row_sums(1:9, 9:1, c(1, 0, NA))

\dontrun{
row_sums(GEHTRAINING, HILFSMITTEL, MEDIKATION, STURZRISIKO)
}

}
