% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-year.R
\name{to_year}
\alias{to_year}
\title{Extracts the year out of a date-like object}
\usage{
to_year(x)
}
\arguments{
\item{x}{a vector coercible to a character vector}
}
\value{
An integer vector of years
}
\description{
Supports \code{mm.yyyy}, \code{[1-4]{1}/yyyy}, \code{yyyy} and uses
\code{\link{as.Date}} to detect other date-like formats.
}
\examples{
to_year("03.1999")

to_year("4/2013")

to_year("2015")

to_year("2019-03-01")

}
